Add `getDiffForCommit` function in `simple-git` library
This commit introduces a new function `getDiffForCommit` in the `simple-git` library. This function fetches the diff for a given commit ID. The function takes in two parameters, the commit ID and an object containing `git` and `logger`. It returns a promise that resolves to the diff of the commit. If there's an error while fetching the diff, it throws an error with a message detailing the commit ID and the error message.