Refactor `getDiffForCommit` function in `simple-git` library
Removed unused `Logger` import and parameter from `getDiffForCommit` function in `simple-git` library. This simplifies the function signature and removes unnecessary dependencies. The function now only takes in a `commitId` and a `git` object of type `SimpleGit`. The function continues to return a promise that resolves to the diff of the commit.