Refactor `index.ts` to simplify `yargs` setup
The `index.ts` file has been refactored to simplify the setup of `yargs`. The verbose setup has been replaced with a more concise usage method. The `commit` and `loadConfig` imports have been adjusted, and the `commit` command setup has been streamlined. A TODO comment has been added to address a type issue on the builder in the future.