Refactor changelog handler to handle dynamic range
In the `handler.ts` file of the changelog command, the hardcoded commit range has been removed. Instead, the range is now dynamically derived from the provided options. The range is expected to be in the format `<from>:<to>`. If it's not, an error message is logged and the process exits. Furthermore, the `getCommitLogRange` function has been updated to accept the `from` and `to` parameters directly, along with an options object. The `noMerges` option has been added to exclude merge commits. The logging of the prompt options has also been removed.