Add `defaultBranch` to config and update schema
Added a new `defaultBranch` property to the config object in `loadConfig.ts` and `types.ts`. This property allows for the specification of the default git branch for the repository. Also, added a TODO comment in `project.ts` to consider adding validation based on JSON schema. Finally, updated `schema.json` to include the new `defaultBranch` property.