Add label to logResult function
Updated the `logResult` function in `src/lib/ui` to accept a `label` parameter. This `label` is now included in the output, providing more context to the log output. Also updated the `generateAndReviewLoop` function to pass the `label` when calling `logResult`. This change enhances the readability and usefulness of the logs by providing more specific information about the results.