Update npm package configuration and build scripts
This commit includes several changes to the npm package configuration and build scripts.

In `.npmignore`, the `src/` and `coverage/` directories, `dist/dst` type definition, and `tsconfig.json` and `README.md` root files were added to the ignore list.

In `package.json`, the `dist` directory was included in the 'files' array, while `!dist/dts` was added. The 'types' property was set to `dist/index.d.ts`. Several scripts were added or modified, including `clean`, `format`, `test`, `test:publish`, `test:jest`, and `prepublishOnly`. The `rollup-plugin-dts` was added to the dependencies.

In `rollup.config.mjs`, the `dts` plugin was imported and the configuration was significantly modified. A new constant `enableBenchmarking` was introduced and used in multiple places. The `config` array was exported instead of an object.

In `tsconfig.json`, the `declaration` property was moved and the `declarationDir` property was changed to `dist/dts`.