Replace `GPT3Tokenizer` with `TokenCounter` in utils
Replaced `GPT3Tokenizer` with `TokenCounter` in `collectDiffs.ts` and `summarizeDiffs.ts` files. This change affects the `collectDiffs`, `summarizeDirectoryDiff` and `SummarizeDiffsOptions` functions which now use `TokenCounter` for token counting.

---

Update import statements and function calls in commit handler

Removed `getTokenizer` and added `getLlm`, `getModelFromService`, and `getTokenCounter` in the import section of `handler.ts`. Also updated the function calls within the `handler` function to use these new imports.

---

Replace `getChain` with `getSummarizationChain` in default parser

In `index.ts`, replaced `getChain` with `getSummarizationChain` in the import statement and in the `fileChangeParser` function. Also renamed the `model` parameter to `llm` in the `options` object.

---

Update import and function calls in changelog handler

Replaced `getModel` with `getLlm` in the import statement and updated the `getApiKeyForModel` and `getLlm` function calls in the `handler` function to use `options.service` instead of `options.model`.

---

Add `tokenizer.ts` and delete `getTokenizer.ts` in utils

Deleted `getTokenizer.ts` and added `tokenizer.ts` which includes the `getTikToken`, `getTokenCount`, and `getTokenCounter` functions.

---

Update types in lib

In `types.ts`, replaced `getTokenizer` and `getModel` with `TokenCounter` and `getLlm` respectively in the `BaseParserOptions` interface.

---

Update dependencies in package.json

Removed `gpt3-tokenizer` and added `tiktoken` in the dependencies of `package.json`.