Add clipboard functionality to changelog handler
This update includes the addition of the `clipboardy` dependency and the implementation of its `writeSync` function in the `handler` function of the changelog command. This allows the result of the changelog command to be automatically copied to the clipboard for easier use. The `logger.log` function is also used to notify the user that the result has been copied to the clipboard. The `clipboardy` dependency has been added to the `package.json` file.