Refactor commit command handler and options
Refactor the `handler` function in the commit command to use the `CommandHandler` type and receive a logger as an argument. This eliminates the need to create a new `Logger` instance within the function. Also, add the `CommitArgv` type to the commit options to better represent the arguments received by the `handler` function. Lastly, update the commit command's index file to use the `commandExecutor` utility when executing the `handler` function.