Refactor changelog command handler and options
In this update, the `handler` function in the `changelog` command has been refactored. The function now accepts two parameters, `argv` and `logger`, instead of just `argv`. This change removes the need to instantiate a new `Logger` within the function.

Additionally, the `ChangelogOptions` type has been expanded to include a new `ChangelogArgv` type. This new type is used in the refactored `handler` function.

The `handler` function in `changelog/index.ts` now uses a `commandExecutor` to run the `handler` from `changelog/handler.ts`.

Lastly, the `LOGO` constant from `ui/helpers` is now logged if the `INTERACTIVE` variable is true.