Update eslint and typescript-eslint versions
This commit updates the versions of `eslint` and `@typescript-eslint/eslint-plugin` and `@typescript-eslint/parser` in the `devDependencies` section of the `package.json` file. The `@rollup/plugin-eslint` version is also updated. This ensures that the project is using the latest versions of these packages, which may include important bug fixes and performance improvements.