Update README with new `coco` features and usage instructions
This update to the README includes a redefinition of `coco` as not just a scribe for git commit messages, but also as a suite of tools to streamline git workflow. The update introduces new commands such as `commit`, `changelog`, and `init`. It also provides detailed instructions on how to get started with `coco`, how to use the different commands, and how to configure `coco` according to user preferences. The section on ignoring files and the configuration documentation have been moved to the wiki for more detailed information.