Refactor config services and update project configuration
This update involves refactoring and updating several files in the `src/lib/config/services` and `src/commands/init` directories. In `services`, the hard-coded start and end comments in `appendToEnvFile` and `appendToGitConfig` functions have been replaced with constants. The `service` value in `defaultConfig` is updated, and the `loadProjectConfig` function is renamed and updated with a file existence check. In `init`, the `appendToProjectConfig` function is renamed, and the `createProjectFileAndReturnPath` function is replaced. A new variable is introduced, and conditions to append to `configFilePath` have been updated. Finally, package installation check is added after configuration write.

In `src/lib/config`, the `DEFAULT_CONFIG` object's `service` property was updated and two new constants were added in `constants.ts`. The `loadProjectConfig` import was replaced in `loadConfig.ts`.

In `src/lib/utils`, the `createProjectFileAndReturnPath.ts` file was deleted and replaced by `getProjectConfigFilePath.ts`, which includes an exported function to get the project configuration file path.