Update tokenizer function to use new openai-types model
This commit updates the `tokenizer` function to use a new `TiktokenModel` from the `openai-types` library, instead of the previous `ServiceModel`. The update is made to support a new feature that requires the use of this new model. The change includes modifying the import statement for the `BPE_Tokenizer` and adding a new export for the `getTokenCounter` function.