Add `jest --watch` script to `package.json`
Added a new script to the `package.json` file to allow `jest` tests to run in watch mode. This allows developers to run tests continuously as changes are made, improving the feedback loop and efficiency of the development process. Simply use the command `npm run test:jest:watch` to activate this feature.